container/requirements.txt
PyJWT==1.6.1
PyNaCl==1.1.2
PySocks==1.6.7
pyyaml==5.4
#pygments==2.7.4
Whoosh==2.7.4
adal==0.6.0
aexpect==1.4.0
aliyun-python-sdk-core==2.13.16
aliyun-python-sdk-ecs==4.19.4
appdirs==1.4.3
applicationinsights==0.11.3
argcomplete==1.9.4
asn1crypto==0.22.0
avocado-framework-plugin-varianter-yaml-to-mux==61.0
avocado-framework==61.0
azure-batch==4.1.3
azure-cli-acr==2.0.24
azure-cli-acs==2.0.32
azure-cli-advisor==0.5.1
azure-cli-ams==0.1.0
azure-cli-appservice==0.1.32
azure-cli-backup==1.1.1
azure-cli-batch==3.2.2
azure-cli-batchai==0.2.2
azure-cli-billing==0.1.8
azure-cli-cdn==0.0.14
azure-cli-cloud==2.0.13
azure-cli-cognitiveservices==0.1.13
azure-cli-command-modules-nspkg==2.0.1
azure-cli-configure==2.0.15
azure-cli-consumption==0.3.1
azure-cli-container==0.1.23
azure-cli-core==2.0.32
azure-cli-cosmosdb==0.1.21
azure-cli-dla==0.0.19
azure-cli-dls==0.0.21
azure-cli-dms==0.0.1
azure-cli-eventgrid==0.1.12
azure-cli-eventhubs==0.1.3
azure-cli-extension==0.0.13
azure-cli-feedback==2.1.1
azure-cli-find==0.2.9
azure-cli-interactive==0.3.20
azure-cli-iot==0.1.20
azure-cli-keyvault==2.0.21
azure-cli-lab==0.0.22
azure-cli-monitor==0.1.6
azure-cli-network==2.1.0
azure-cli-nspkg==3.0.2
azure-cli-profile==2.0.24
azure-cli-rdbms==0.2.2
azure-cli-redis==0.2.13
azure-cli-reservations==0.1.2
azure-cli-resource==2.0.28
azure-cli-role==2.0.23
azure-cli-servicebus==0.1.2
azure-cli-servicefabric==0.0.12
azure-cli-sql==2.0.25
azure-cli-storage==2.0.32
azure-cli-vm==2.0.31
azure-cli==2.0.32
azure-common==1.1.11
azure-datalake-store==0.0.19
azure-graphrbac==0.40.0
azure-keyvault==0.3.7
azure-mgmt-advisor==1.0.1
azure-mgmt-authorization==0.40.0
azure-mgmt-batch==4.1.0
azure-mgmt-batchai==1.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==1.0.0
azure-mgmt-cognitiveservices==1.0.0
azure-mgmt-compute==4.0.0rc2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==0.4.0
azure-mgmt-containerregistry==2.0.0
azure-mgmt-containerservice==3.0.1
azure-mgmt-cosmosdb==0.4.0
azure-mgmt-datalake-analytics==0.2.0
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.2.0
azure-mgmt-datamigration==0.1.0
azure-mgmt-devtestlabs==2.0.0
azure-mgmt-dns==2.0.0rc1
azure-mgmt-eventgrid==0.4.0
azure-mgmt-eventhub==1.2.0
azure-mgmt-iothub==0.4.0
azure-mgmt-iothubprovisioningservices==0.1.0
azure-mgmt-keyvault==0.40.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0rc1
azure-mgmt-monitor==0.5.0
azure-mgmt-msi==0.1.0
azure-mgmt-network==2.0.0rc2
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==1.1.0
azure-mgmt-recoveryservices==0.1.0
azure-mgmt-recoveryservicesbackup==0.1.1
azure-mgmt-redis==4.1.0
azure-mgmt-reservations==0.1.0
azure-mgmt-resource==1.2.1
azure-mgmt-servicebus==0.4.0
azure-mgmt-servicefabric==0.1.0
azure-mgmt-sql==0.8.6
azure-mgmt-storage==1.5.0
azure-mgmt-trafficmanager==0.40.0
azure-mgmt-web==0.35.0
azure-multiapi-storage==0.2.0
azure-nspkg==2.0.0
azure-storage-blob==1.1.0
azure-storage-common==1.1.0
azure-storage-nspkg==3.0.0
bcrypt==3.1.3
boto3==1.7.23
botocore==1.10.23
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.4
colorama==0.3.3
cryptography==3.3.2
decorator==4.3.0
deprecation==2.0.2
docutils==0.14
dogpile.cache==0.6.5
enum34==1.1.6
funcsigs==1.0.2
futures==3.1.1
humanfriendly==4.12.1
idna==2.5
ipaddress==1.0.18
iso8601==0.1.12
isodate==0.6.0
jmespath==0.7.1
jsonpatch==1.23
jsonpointer==2.0
keystoneauth1==3.6.2
knack==0.3.2
mock==2.0.0
monotonic==1.5
msrest==0.4.29
msrestazure==0.4.31
munch==2.3.2
netifaces==0.10.7
oauthlib==2.0.7
openstacksdk==0.13.0
os-service-types==1.2.0
packaging==17.1
paramiko==2.2.4
pathlib2==2.3.2
pbr==4.0.3
prompt-toolkit==1.0.15
pyOpenSSL==18.0.0
pyasn1==0.2.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.6
pydocumentdb==2.3.2
pyparsing==2.1.10
python-dateutil==2.6.1
pytz==2016.6.1
requests-oauthlib==0.8.0
requests==2.20.0
requestsexceptions==1.4.0
s3transfer==0.1.10
scandir==1.7
scp==0.11.0
six==1.10.0
sshtunnel==0.1.3
stevedore==1.28.0
tabulate==0.8.2
urllib3==1.26.5
vsts-cd-manager==1.0.1
wcwidth==0.1.7
websocket-client==0.47.0
xmltodict==0.11.0
avocado-framework-plugin-result-html==61.0
avocado-framework-plugin-loader-yaml==61.0
azure-batch-extensions==1.1.2
backports.ssl-match-hostname==3.5.0.1
configparser==3.5.0
entrypoints==0.2.3
Fabric==1.14.0
jinja2==2.11.3
keyring==15.0.0
MarkupSafe==1.1.1
pathlib==1.0.1
ply==3.9
pyliblzma==0.5.3
pystache==0.5.4
SecretStorage==2.3.1
simplejson==3.10.0
